+++ src/main/java/teammates/common/util/Const.java
@@ -123,7 +123,6 @@ public final class Const {
 
         public static final String ACCOUNT_REQUEST_ID = "id";
         public static final String ACCOUNT_REQUEST_STATUS = "status";
-        public static final String ACCOUNT_ID = "accountid";
 
         public static final String FEEDBACK_SESSION_NAME = "fsname";
         public static final String FEEDBACK_SESSION_STARTTIME = "starttime";
